+++ CHANGELOG.markdown
@@ -2,7 +2,7 @@
 
 ## Lleidasms 1.1.0
 - Client class features
-  * send SMS text plain and binary (base64) format
+  * send SMS text plain (ASCII), unicode and binary (base64) format
 
 ## Lleidasms 1.0.2